Get the DLL builds going.
